backend/requirements.txt
flask==3.1.0
flask-sqlalchemy==3.1.1
flask-cors==5.0.1
requests==2.32.3
python-dotenv==1.1.0
